Added stuff changed things. I've forgotten.

+++ WaveGame/NavigationMesh.cpp
@@ -28,6 +28,7 @@ void Engine::Core::NavigationNode::DebugDraw(shared_ptr<RenderWindow> Window)
 
 Engine::Core::NavigationMesh::NavigationMesh(shared_ptr<RenderWindow> Window, Map M)
 {
+	///Create a new thread so that it can run in the background and not cause the game to lock up.
 	this->LoadingThread = thread(&NavigationMesh::CreateNavigationMesh, this, Window, M);
 	this->LoadingThread.detach();
 }
@@ -38,12 +39,12 @@ void Engine::Core::NavigationMesh::CreateNavigationMesh(const std::shared_ptr<sf
 
 	auto Props = M.GetProps();
 	
-	for (int x = 0; x < (int)Window->getSize().x; x += NODE_DISTANCE)
+	for (float x = 0; x < (float)Window->getSize().x; x += NODE_DISTANCE)
 	{
-		for (int y = 0; y < (int)Window->getSize().y; y += NODE_DISTANCE)
+		for (float y = 0; y < (float)Window->getSize().y; y += NODE_DISTANCE)
 		{
 			///Used to make sure theres no props in this location.
-			FloatRect Tester = FloatRect(x, y, 1, 1);
+			FloatRect Tester = FloatRect(x, y, 1.0f, 1.0f);
 			///Used for finding if the node is placed in a location where a prop is.
 			bool DoesCollision = false;
 			
@@ -73,7 +74,7 @@ void Engine::Core::NavigationMesh::DebugDraw(shared_ptr<RenderWindow> Window)
 {
 	if (this->LoadingMuxtex.try_lock())
 	{
-		for (int i = 0; i < this->NavNodes.size(); i++)
+		for (size_t i = 0; i < this->NavNodes.size(); i++)
 			this->NavNodes[i].DebugDraw(Window);
 		this->LoadingMuxtex.unlock();
 	}
